Pearson Quizzes
Write a program that asks the user to enter five test scores. The program should display a letter grade for each score and the average test score. Write the following methods in the program:

calcAverage:
This method should accept five test scores as arguments and return the average of the scores.

determineGrade:
This method should accept a test score as an argument and return a letter grade for the score, based on the following grading scale:

Score Letter Grade
90-100 A
80-89 B
70-79 C
60-69 D
Below 60 F

Sample Run
java TestAvgAndGrade

·Enter·test·grade·for·student·1:55↵
·Enter·test·grade·for·student·2:65↵
·Enter·test·grade·for·student·3:75↵
·Enter·test·grade·for·student·4:85↵
·Enter·test·grade·for·student·5:95↵
·The·letter·grades·are·as·follows:↵
Student·1:·F↵
Student·2:·D↵
Student·3:·C↵
Student·4:·B↵
Student·5:·A↵
The·average·grade·was:·75.00↵